+++ src/lilypond.ts
@@ -32,6 +32,27 @@ export const initCompile = () => {
   )
 }
 
+const getCompilingStatusBarItem = () => {
+  const item = vscode.window.createStatusBarItem(
+    vscode.StatusBarAlignment.Left,
+    0
+  )
+  item.text = `$(sync~spin) Compiling...`
+  item.tooltip = `You can kill the compilation process using the \`VSLilyPond: Kill Compilation Process\` command`
+  return item
+}
+
+const showCompilationFailedStatusBarItem = (timeoutMS = 5000) => {
+  const item = vscode.window.createStatusBarItem(
+    vscode.StatusBarAlignment.Left,
+    0
+  )
+  item.text = `$(x) Compilation Failed`
+  item.tooltip = `See the \`VSLilyPond: Compilation\` output for more information`
+  item.show()
+  setTimeout(() => item.hide(), timeoutMS)
+}
+
 const outputToChannel = async (msg: string, show = false) => {
   if (compileOutputChannel) {
     compileOutputChannel.appendLine(msg)
@@ -46,7 +67,6 @@ const outputToChannel = async (msg: string, show = false) => {
     )
   }
 }
-
 const getCompilationFilePath = (
   compileMode: CompileMode,
   activeTextDocument: vscode.TextDocument
@@ -115,32 +135,11 @@ export const compile = async (
   mute = false,
   textDocument: vscode.TextDocument | undefined = undefined
 ) => {
+  const compilingStatasBarItem = getCompilingStatusBarItem()
+  compilingStatasBarItem.show()
+
   try {
     if (compileProcess) {
-      // basically, onSave compile cannot override onCompile/onCompileSpecific compile
-      if (
-        compileMode === CompileMode.onSave &&
-        (compileProcess.compileMode === CompileMode.onCompile ||
-          compileProcess.compileMode === CompileMode.onCompileSpecific)
-      ) {
-        // currently running compile process is onCompile and the current caller is onSave
-        // return and do nothing
-        return
-      }
-
-      // do not let onCompile/onCompileSpecific override each other
-      if (
-        compileProcess.compileMode === CompileMode.onCompile ||
-        compileProcess.compileMode === CompileMode.onCompileSpecific
-      ) {
-        logger(
-          `Cannot compile - there is an existing compilation, please wait for that to finish.`,
-          LogLevel.error,
-          false
-        )
-        return
-      }
-
       killCompilation(true)
     }
 
@@ -195,14 +194,22 @@ export const compile = async (
       if (code === 0) {
         logger(`Compilation successful`, LogLevel.info, mute)
         outputToChannel(`Compilation successful`)
+      } else if (code === null) {
+        // here, the compilation process is replaced (i.e. killed above)
+        logger(`Compilation killed`, LogLevel.error, mute)
+        outputToChannel(`Compilation killed`, false)
       } else {
         logger(`Compilation failed`, LogLevel.error, mute)
-        outputToChannel(`Compilation failed`, true)
+        outputToChannel(`Compilation failed`)
+        showCompilationFailedStatusBarItem()
       }
       compileProcess = undefined
+      compilingStatasBarItem.hide()
     })
   } catch (err) {
     logger(err.message, LogLevel.error, mute)
     outputToChannel(`Compilation failed: ${err.message}`, true)
+    showCompilationFailedStatusBarItem()
+    compilingStatasBarItem.hide()
   }
 }