+++ src/midi-out.ts
@@ -153,7 +153,7 @@ export namespace MIDIOut {
 
     if (!midiFileName) {
       throw new Error(
-        `Cannot find MIDI file to play - make sure you are outputting a MIDI file`
+        `Cannot find MIDI file to play - make sure you are outputting a MIDI file, and the editor is currently active on the file that produces the MIDI.`
       )
     }
     return midiFileName
@@ -171,6 +171,10 @@ export namespace MIDIOut {
       MIDIOutState.currMidiFilePath = midiFileName
       MIDIOutState.player = smf.player()
 
+      const outputs = await getOutputMIDIDevices()
+      if (outputs.length === 0) {
+        throw new Error(`No output MIDI devices are found.`)
+      }
       const config = getConfiguration()
       const midiout = config.midiPlayback.output.length
         ? JZZ().openMidiOut(config.midiPlayback.output)
@@ -213,9 +217,7 @@ export namespace MIDIOut {
 
   export const playMIDI = async () => {
     try {
-      resetMIDI(true)
       await loadMIDI()
-
       if (MIDIOutState.player) {
         MIDIOutState.player.play()
         MIDIOutState.playing = true
@@ -231,9 +233,7 @@ export namespace MIDIOut {
 
   export const playMIDIFrom = async () => {
     try {
-      resetMIDI(true)
       await loadMIDI()
-
       if (MIDIOutState.player) {
         // get the maximum duration and ask user to input the required duration
         const durationMS = MIDIOutState.player.durationMS()
@@ -393,12 +393,17 @@ export namespace MIDIOut {
     }
   }
 
-  // set output midi device
-  export const setOutputMIDIDevice = async () => {
+  const getOutputMIDIDevices = async () => {
     const outputs: string[] = JZZ()
       .info()
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       .outputs.map((x: any) => x.name)
+    return outputs
+  }
+
+  // set output midi device
+  export const setOutputMIDIDevice = async () => {
+    const outputs = await getOutputMIDIDevices()
     vscode.window.showQuickPick(outputs).then((val: string | undefined) => {
       if (val) {
         const config = getConfiguration()
@@ -407,25 +412,9 @@ export namespace MIDIOut {
     })
   }
 
-  const MIDIFileExists = async (): Promise<boolean> => {
-    try {
-      const midiFilePath = await getMidiFilePathFromWindow()
-      if (midiFilePath) {
-        return true
-      }
-    } catch (err) {
-      return false
-    }
-    return false
-  }
-
   const shouldShowStatusBarItems = async (): Promise<boolean> => {
     const activeTextEditor = vscode.window.activeTextEditor
-    if (
-      activeTextEditor &&
-      activeTextEditor.document.languageId === langId &&
-      (await MIDIFileExists())
-    ) {
+    if (activeTextEditor && activeTextEditor.document.languageId === langId) {
       return true
     }
     return false